backend/pyproject.toml
[tool.poetry]
name = "TuneScript"
version = "0.1.0"
description = "Piano notes transcription application"
authors = ["Michał Jagoda <michal-jagoda@o2.pl>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aniso8601 = "9.0.1"
asgiref = "3.8.1"
build = "1.2.1"
CacheControl = "0.14.0"
certifi = "2024.6.2"
cffi = "1.16.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
cryptography = "42.0.8"
distlib = "0.3.8"
Django = "5.0.6"
django-allauth = "0.63.3"
django-graphql-jwt = "0.4.0"
django-rest-auth = "0.9.5"
djangorestframework = "3.15.1"
dnspython = "2.6.1"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.14.0"
graphene = "3.3"
graphene-django = "3.2.1"
graphene-file-upload = "1.3.0"
graphql-core = "3.2.3"
graphql-relay = "3.2.0"
idna = "3.7"
importlib_metadata = "7.1.0"
installer = "0.7.0"
jeepney = "0.8.0"
keyring = "24.3.1"
more-itertools = "10.3.0"
msgpack = "1.0.8"
mysqlclient = "2.2.4"
numpy = "1.26.4"
packaging = "24.1"
pexpect = "4.9.0"
pillow = "10.3.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
promise = "2.3"
ptyprocess = "0.7.0"
pycparser = "2.22"
PyJWT = "2.8.0"
pyproject_hooks = "1.1.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
rapidfuzz = "3.9.3"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
SecretStorage = "3.3.3"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.5.0"
text-unidecode = "1.3"
tomlkit = "0.12.5"
trove-classifiers = "2024.5.22"
typing_extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.1"
virtualenv = "20.26.2"
zipp = "3.19.2"
piano-transcription-inference = "^0.0.5"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.5"
torch = "^1.10.0"

[tool.poetry.dev-dependencies]
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
